Support a usecase without signin (#44)
Fixes #43

The user is now able to use the basic editor functionality without signin except for:

- AI agent features
- Firebase integrations

+++ core/platform/src/desktopMain/kotlin/io/composeflow/platform/GetOrCreateCloudProjectSaver.desktop.kt
@@ -3,6 +3,7 @@ package io.composeflow.platform
 import co.touchlab.kermit.Logger
 import com.github.michaelbull.result.mapBoth
 import io.composeflow.cloud.storage.GoogleCloudStorageWrapper
+import io.composeflow.datastore.ANONYMOUSE_USER_ID
 import io.composeflow.datastore.PROJECT_YAML_FILE_NAME
 import io.composeflow.datastore.ProjectSaver
 import io.composeflow.datastore.ProjectYamlNameWithLastModified
@@ -27,34 +28,38 @@ class CloudProjectSaverImpl(
         yamlContent: String,
         syncWithCloud: Boolean,
     ) = withContext(ioDispatcher) {
-        cloudStorageWrapper
-            .uploadFile(
-                userId = userId,
-                projectId = projectId,
-                fileName = PROJECT_YAML_FILE_NAME,
-                content = yamlContent,
-            ).mapBoth(
-                success = {
-                    // Do nothing
-                },
-                failure = {
-                    Logger.e("Failed to save the project yaml", it)
-                },
-            )
+        if (userId != ANONYMOUSE_USER_ID && syncWithCloud) {
+            cloudStorageWrapper
+                .uploadFile(
+                    userId = userId,
+                    projectId = projectId,
+                    fileName = PROJECT_YAML_FILE_NAME,
+                    content = yamlContent,
+                ).mapBoth(
+                    success = {
+                        // Do nothing
+                    },
+                    failure = {
+                        Logger.e("Failed to save the project yaml", it)
+                    },
+                )
+        }
     }
 
     override suspend fun deleteProject(
         userId: String,
         projectId: String,
     ) = withContext(ioDispatcher) {
-        cloudStorageWrapper.deleteFolder("$userId/$projectId/").mapBoth(
-            success = {
-                // Do nothing
-            },
-            failure = {
-                Logger.e("Failed to delete the project yaml", it)
-            },
-        )
+        if (userId != ANONYMOUSE_USER_ID) {
+            cloudStorageWrapper.deleteFolder("$userId/$projectId/").mapBoth(
+                success = {
+                    // Do nothing
+                },
+                failure = {
+                    Logger.e("Failed to delete the project yaml", it)
+                },
+            )
+        }
     }
 
     @OptIn(kotlin.time.ExperimentalTime::class)
@@ -62,36 +67,44 @@ class CloudProjectSaverImpl(
         userId: String,
         projectId: String,
     ): ProjectYamlNameWithLastModified? =
-        withContext(ioDispatcher) {
-            cloudStorageWrapper.getFile("$userId/$projectId/$PROJECT_YAML_FILE_NAME").mapBoth(
-                success = {
-                    it.contentBytes?.let { contentBytes ->
-                        ProjectYamlNameWithLastModified(
-                            contentBytes.toString(Charsets.UTF_8),
-                            it.updateTime,
-                        )
-                    }
-                },
-                failure = {
-                    Logger.e("Failed to load project", it)
-                    null
-                },
-            )
+        if (userId != ANONYMOUSE_USER_ID) {
+            withContext(ioDispatcher) {
+                cloudStorageWrapper.getFile("$userId/$projectId/$PROJECT_YAML_FILE_NAME").mapBoth(
+                    success = {
+                        it.contentBytes?.let { contentBytes ->
+                            ProjectYamlNameWithLastModified(
+                                contentBytes.toString(Charsets.UTF_8),
+                                it.updateTime,
+                            )
+                        }
+                    },
+                    failure = {
+                        Logger.e("Failed to load project", it)
+                        null
+                    },
+                )
+            }
+        } else {
+            null
         }
 
     override suspend fun loadProjectIdList(userId: String): List<String> =
-        withContext(ioDispatcher) {
-            cloudStorageWrapper.listFile(userId).mapBoth(
-                success = { blobList ->
-                    blobList.map {
-                        it.folderName
-                    }
-                },
-                failure = {
-                    Logger.e("Failed to load projectId list", it)
-                    emptyList()
-                },
-            )
+        if (userId != ANONYMOUSE_USER_ID) {
+            withContext(ioDispatcher) {
+                cloudStorageWrapper.listFile(userId).mapBoth(
+                    success = { blobList ->
+                        blobList.map {
+                            it.folderName
+                        }
+                    },
+                    failure = {
+                        Logger.e("Failed to load projectId list", it)
+                        emptyList()
+                    },
+                )
+            }
+        } else {
+            emptyList()
         }
 
     override suspend fun deleteCacheProjects() {
